reorganize some files
