+++ Saule/Queries/Pagination/PaginationQuery.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 
-namespace Saule.Queries
+namespace Saule.Queries.Pagination
 {
     internal class PaginationQuery
     {
