+++ Saule/Queries/Query.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using Saule.Queries.Pagination;
 
 namespace Saule.Queries
 {
