+++ Saule/Queries/Sorting/SortingContext.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace Saule.Queries.Sorting
+{
+    internal class SortingContext
+    {
+        public IEnumerable<string> Properties { get; }
+    }
+}
