+++ Saule/Saule.csproj
@@ -77,6 +77,7 @@
     <Compile Include="JsonApiSerializerOfT.cs" />
     <Compile Include="Queries\Query.cs" />
     <Compile Include="Queries\QueryContext.cs" />
+    <Compile Include="Queries\Sorting\SortingContext.cs" />
     <Compile Include="ResourceRelationshipOfT.cs" />
     <Compile Include="Serialization\CanonicalUrlPathBuilder.cs" />
     <Compile Include="Serialization\DefaultUrlPathBuilder.cs" />
@@ -85,11 +86,11 @@
     <Compile Include="Http\ReturnsResourceAttribute.cs" />
     <Compile Include="Http\JsonApiMediaTypeFormatter.cs" />
     <Compile Include="RelationshipKind.cs" />
-    <Compile Include="Queries\PaginationContext.cs" />
-    <Compile Include="Queries\PaginationInterpreter.cs" />
+    <Compile Include="Queries\Pagination\PaginationContext.cs" />
+    <Compile Include="Queries\Pagination\PaginationInterpreter.cs" />
     <Compile Include="Queries\QueryableExtensions.cs" />
     <Compile Include="Queries\QueryMethod.cs" />
-    <Compile Include="Queries\PaginationQuery.cs" />
+    <Compile Include="Queries\Pagination\PaginationQuery.cs" />
     <Compile Include="Serialization\ApiError.cs" />
     <Compile Include="ResourceAttribute.cs" />
     <Compile Include="ResourceRelationship.cs" />
