+++ Tests/Serialization/UrlConstructionTests.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Saule.Queries;
 using Saule;
+using Saule.Queries.Pagination;
 using Saule.Serialization;
 using Tests.Helpers;
 using Tests.Models;
